Enhance nullability and type safety across application
This commit introduces several updates to command handlers, services, and DTOs to improve nullability handling and type safety. Key changes include:

- Added null-forgiving operators to command handler responses.
- Introduced new properties and methods in DTOs for better data encapsulation.
- Updated query handlers with additional null checks for robustness.
- Refactored validation logic in command validators for conditional rules.
- Enhanced handling of portal and incident metrics with new enums and value objects.
- Improved caching mechanisms with new methods for cache invalidation.
- Updated health check and notification services for better error handling.
- Refined repository interfaces to support detailed portal information and batch operations.

These changes aim to enhance maintainability, readability, and performance while adhering to best practices in C# development.

+++ apps/api/CentralCommand.Api/Services/MetricsCalculationService.cs
@@ -89,7 +89,7 @@ public class MetricsCalculationService : IMetricsCalculationService
         trends["ResponseTimeTrend"] = CalculateTrend(sortedHistory.Select(h => h.ResponseTime).ToList());
         trends["UptimeTrend"] = CalculateTrend(sortedHistory.Select(h => h.Uptime).ToList());
         trends["ErrorRateTrend"] = CalculateTrend(sortedHistory.Select(h => h.ErrorRate).ToList());
-        trends["RequestsPerMinuteTrend"] = CalculateTrend(sortedHistory.Select(h => h.RequestsPerMinute).ToList());
+        trends["RequestsPerMinuteTrend"] = CalculateTrend(sortedHistory.Select(h => (double)h.RequestsPerMinute).ToList());
 
         return trends;
     }
